requirements2.txt
streamlit
pandas
google-cloud-firestore
firebase-admin
